drop 32-bit windows, add a little more testing
32-bit build stopped working. Probably salvagable but the cost in
time of keeping it going is significant. Going to drop support and
see who complains. Usage of 32-bit windows has been dwindling for
years.

In passing, make the tiny smoke test also exercise the part of Grist
that reads a document, to exercise node-sqlite3, which is the most
delicate part of the packaging historically.